pull from upstream

+++ components/camel-yammer/src/main/java/org/apache/camel/component/yammer/YammerMessagePollingConsumer.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
@@ -75,7 +75,7 @@ public class YammerMessagePollingConsumer extends ScheduledPollConsumer {
             args.append(limit);
         }        
 
-        int olderThan = endpoint.getConfig().getOlderThan();
+        long olderThan = endpoint.getConfig().getOlderThan();
         if (olderThan > 0) {
             if (args.length() > 0) {
                 args.append("&");
@@ -84,7 +84,7 @@ public class YammerMessagePollingConsumer extends ScheduledPollConsumer {
             args.append(olderThan);
         }        
 
-        int newerThan = endpoint.getConfig().getNewerThan();
+        long newerThan = endpoint.getConfig().getNewerThan();
         if (newerThan > 0) {
             if (args.length() > 0) {
                 args.append("&");
